Moved all options into specific tabs
