+++ src/blocks/custom/form/components/form-email-options.js
@@ -1,0 +1,18 @@
+import { __ } from '@wordpress/i18n';
+import { Fragment } from '@wordpress/element';
+import { SelectControl } from '@wordpress/components';
+
+export const FormEmailOptions = (props) => {
+  const {
+    type,
+    crmEntitiesAsOptions,
+    dynamicsEntity,
+    isDynamicsCrmUsed,
+    onChangeDynamicsEntity,
+  } = props;
+
+  return (
+    <Fragment>
+    </Fragment>
+  );
+};
