+++ src/blocks/custom/form/components/form-options.js
@@ -1,24 +1,9 @@
 import { __ } from '@wordpress/i18n';
 import { Fragment } from '@wordpress/element';
-import { PanelBody, TextControl, SelectControl, BaseControl, TabPanel, Dashicon } from '@wordpress/components';
-import { RichText } from '@wordpress/block-editor';
-
-/**
- * Custom action which changes the "theme" attribute for this block and all it's innerBlocks.
- *
- * @param {string} newTheme New value for theme attribute
- * @param {function} onChangeTheme Prebuilt action for form block.
- */
-const updateThemeForAllInnerBlocks = (newTheme, onChangeTheme) => {
-  const thisBlock = wp.data.select('core/block-editor').getSelectedBlock();
-  if (thisBlock.innerBlocks) {
-    thisBlock.innerBlocks.forEach((innerBlock) => {
-      innerBlock.attributes.theme = newTheme;
-      wp.data.dispatch('core/block-editor').updateBlock(innerBlock.clientId, innerBlock);
-    });
-  }
-  onChangeTheme(newTheme);
-};
+import { PanelBody, TextControl, TabPanel, Dashicon } from '@wordpress/components';
+import { FormGeneralOptions } from './form-general-options';
+import { FormDynamicsCrmOptions } from './form-dynamics-crm-options';
+import { FormEmailOptions } from './form-email-options';
 
 export const FormOptions = (props) => {
   const {
@@ -65,7 +50,6 @@ export const FormOptions = (props) => {
 
   const themeAsOptions = hasThemes ? themes.map((tempTheme) => ({ label: tempTheme, value: tempTheme })) : [];
 
-  // All Dynamics CRM config stuff
   let crmEntitiesAsOptions = [];
   if (isDynamicsCrmUsed) {
     crmEntitiesAsOptions = dynamicsCrm.availableEntities.map((entity) => ({ label: entity, value: entity }));
@@ -104,54 +88,20 @@ export const FormOptions = (props) => {
                 <strong className="notice-title">{__('General Options', 'eightshift-forms')}</strong>
                 <p>{__('These are general form options.', 'eightshift-forms')}</p>
                 <br />
-                {onChangeType &&
-                  <SelectControl
-                    label={__('Type', 'eightshift-forms')}
-                    value={type}
-                    help={__('Choose what will this form do on submit', 'eightshift-forms')}
-                    options={formTypes}
-                    onChange={onChangeType}
-                  />
-                }
-                {onChangeTheme && hasThemes &&
-                  <SelectControl
-                    label={__('Theme', 'eightshift-forms')}
-                    help={__('Choose your form theme.', 'eightshift-forms')}
-                    value={theme}
-                    options={themeAsOptions}
-                    onChange={(newTheme) => {
-                      updateThemeForAllInnerBlocks(newTheme, onChangeTheme);
-                    }}
-                  />
-                }
-
-                {onChangeSuccessMessage &&
-                  <BaseControl
-                    label={__('Success message', 'eightshift-forms')}
-                    help={__('Message that the user will see if forms successfully submits.', 'eightshift-forms')}
-                  >
-                    <RichText
-                      className={richTextClass}
-                      placeholder={__('Add your success message', 'eightshift-forms')}
-                      onChange={onChangeSuccessMessage}
-                      value={successMessage}
-                    />
-                  </BaseControl>
-                }
-
-                {onChangeErrorMessage &&
-                  <BaseControl
-                    label={__('Error message', 'eightshift-forms')}
-                    help={__('Message that the user will see if forms fails to submit for whatever reason.', 'eightshift-forms')}
-                  >
-                    <RichText
-                      className={richTextClass}
-                      placeholder={__('Add your error message', 'eightshift-forms')}
-                      onChange={onChangeErrorMessage}
-                      value={errorMessage}
-                    />
-                  </BaseControl>
-                }
+                <FormGeneralOptions
+                  type={type}
+                  formTypes={formTypes}
+                  theme={theme}
+                  themeAsOptions={themeAsOptions}
+                  hasThemes={hasThemes}
+                  richTextClass={richTextClass}
+                  successMessage={successMessage}
+                  errorMessage={errorMessage}
+                  onChangeType={onChangeType}
+                  onChangeTheme={onChangeTheme}
+                  onChangeSuccessMessage={onChangeSuccessMessage}
+                  onChangeErrorMessage={onChangeErrorMessage}
+                />
               </Fragment>
             )}
             {tab.name === 'email' && (
@@ -160,6 +110,8 @@ export const FormOptions = (props) => {
                 <strong className="notice-title">{__('Email Options', 'eightshift-forms')}</strong>
                 <p>{__('These are options for when your form is sending emails.', 'eightshift-forms')}</p>
                 <br />
+                <FormEmailOptions
+                />
               </Fragment>
             )}
             {tab.name === 'dynamics-crm' && (
@@ -168,15 +120,13 @@ export const FormOptions = (props) => {
                 <strong className="notice-title">{__('Dynamics CRM Options', 'eightshift-forms')}</strong>
                 <p>{__('These are options for when your form is sending data to Dynamics CRM.', 'eightshift-forms')}</p>
                 <br />
-                {onChangeDynamicsEntity && isDynamicsCrmUsed && type === 'dynamics-crm' &&
-                  <SelectControl
-                    label={__('CRM Entity', 'eightshift-forms')}
-                    help={__('Please enter the name of the entity record to which you wish to add records.', 'eightshift-forms')}
-                    value={dynamicsEntity}
-                    options={crmEntitiesAsOptions}
-                    onChange={onChangeDynamicsEntity}
-                  />
-                }
+                <FormDynamicsCrmOptions
+                  type={type}
+                  crmEntitiesAsOptions={crmEntitiesAsOptions}
+                  dynamicsEntity={dynamicsEntity}
+                  isDynamicsCrmUsed={isDynamicsCrmUsed}
+                  onChangeDynamicsEntity={onChangeDynamicsEntity}
+                />
               </Fragment>
             )}
           </Fragment>
